+++ CSharpRepl/Program.cs
@@ -9,9 +9,11 @@ using System.IO;
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
+using CSharpRepl.Commands;
 using CSharpRepl.InjectedHook.Contracts;
 using CSharpRepl.Logging;
 using CSharpRepl.PrettyPromptConfig;
+using CSharpRepl.Repls;
 using CSharpRepl.Services;
 using CSharpRepl.Services.Logging;
 using CSharpRepl.Services.Remote;
