+++ Tests/CSharpRepl.Tests/ConfigurationFileTests.cs
@@ -2,6 +2,7 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
+using CSharpRepl.Commands;
 using System.IO;
 using Xunit;
